+++ aquadoggo/src/db/provider.rs
@@ -1,14 +1,14 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 use async_trait::async_trait;
+use p2panda_rs::document::{DocumentId, DocumentViewId};
+use p2panda_rs::hash::Hash;
+use p2panda_rs::schema::SchemaId;
+use p2panda_rs::storage_provider::errors::OperationStorageError;
+use p2panda_rs::storage_provider::traits::StorageProvider;
 use sqlx::query_scalar;
 
-use p2panda_rs::document::DocumentId;
-use p2panda_rs::hash::Hash;
-use p2panda_rs::storage_provider::traits::StorageProvider;
-
-use crate::db::stores::StorageEntry;
-use crate::db::stores::StorageLog;
+use crate::db::stores::{StorageEntry, StorageLog};
 use crate::db::Pool;
 use crate::errors::StorageProviderResult;
 use crate::graphql::client::{
@@ -71,5 +71,120 @@ impl StorageProvider<StorageEntry, StorageLog> for SqlStorage {
         });
 
         Ok(hash)
+    }
+}
+
+impl SqlStorage {
+    /// Returns the schema id for a document view.
+    ///
+    /// Returns `None` if this document view is not found.
+    pub async fn get_schema_by_document_view(
+        &self,
+        view_id: &DocumentViewId,
+    ) -> StorageProviderResult<Option<SchemaId>> {
+        let result: Option<String> = query_scalar(
+            "
+            SELECT
+                schema_id
+            FROM
+                document_views
+            WHERE
+                document_view_id = $1
+            ",
+        )
+        .bind(view_id.as_str())
+        .fetch_optional(&self.pool)
+        .await
+        .map_err(|e| OperationStorageError::FatalStorageError(e.to_string()))?;
+
+        // Unwrap because we expect no invalid schema ids in the db.
+        Ok(result.map(|id_str| id_str.parse().unwrap()))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::convert::TryFrom;
+    use std::str::FromStr;
+
+    use p2panda_rs::document::{DocumentView, DocumentViewFields, DocumentViewId};
+    use p2panda_rs::entry::{LogId, SeqNum};
+    use p2panda_rs::identity::Author;
+    use p2panda_rs::operation::{AsOperation, OperationId};
+    use p2panda_rs::schema::SchemaId;
+    use p2panda_rs::storage_provider::traits::{AsStorageEntry, EntryStore};
+    use p2panda_rs::test_utils::constants::TEST_SCHEMA_ID;
+    use p2panda_rs::test_utils::fixtures::random_document_view_id;
+    use rstest::rstest;
+
+    use crate::db::stores::test_utils::{test_db, TestSqlStore};
+    use crate::db::traits::DocumentStore;
+
+    // Get a `DocumentView` that exists in the db.
+    async fn insert_document_view(db: &TestSqlStore) -> DocumentViewId {
+        let author = Author::try_from(db.key_pairs[0].public_key().to_owned()).unwrap();
+        let entry = db
+            .store
+            .get_entry_at_seq_num(&author, &LogId::new(1), &SeqNum::new(1).unwrap())
+            .await
+            .unwrap()
+            .unwrap();
+        let operation_id: OperationId = entry.hash().into();
+        let document_view_id: DocumentViewId = operation_id.clone().into();
+        let document_view = DocumentView::new(
+            &document_view_id,
+            &DocumentViewFields::new_from_operation_fields(
+                &operation_id,
+                &entry.operation().fields().unwrap(),
+            ),
+        );
+        let result = db
+            .store
+            .insert_document_view(&document_view, &SchemaId::from_str(TEST_SCHEMA_ID).unwrap())
+            .await;
+
+        assert!(result.is_ok());
+        document_view_id
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_get_schema_for_view(
+        #[from(test_db)]
+        #[with(1, 1)]
+        #[future]
+        db: TestSqlStore,
+    ) {
+        let db = db.await;
+
+        // Get a `DocumentView` that exists in the db.
+        let document_view_id = insert_document_view(&db).await;
+        let result = db
+            .store
+            .get_schema_by_document_view(&document_view_id)
+            .await;
+
+        assert!(result.is_ok());
+        assert_eq!(result.unwrap().unwrap().name(), "venue");
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_get_schema_for_missing_view(
+        random_document_view_id: DocumentViewId,
+        #[from(test_db)]
+        #[with(1, 1)]
+        #[future]
+        db: TestSqlStore,
+    ) {
+        let db = db.await;
+
+        let result = db
+            .store
+            .get_schema_by_document_view(&random_document_view_id)
+            .await;
+
+        assert!(result.is_ok());
+        assert!(result.unwrap().is_none());
     }
 }