+++ aquadoggo/src/db/stores/mod.rs
@@ -5,6 +5,7 @@ mod entry;
 mod log;
 mod operation;
 mod schema;
+mod task;
 #[cfg(test)]
 pub mod test_utils;